MyProject/requirements.txt
python=3.10
selenium=4.1.0
selenium~=4.1.3
python-dotenv~=0.20.0
pytest~=7.1.2